Fix: Resolve MDX suspense error in client components
